### 2018-08-11 #891 @whedon
[PRE REVIEW]: q2-sample-classifier: machine-learning tools for microbiome classification and regression
**Submitting author:** @nbokulich (<a href="http://orcid.org/0000-0002-1784-8935">Nicholas Bokulich</a>)
**Repository:** <a href="https://github.com/caporaso-lab/sample-classifier-ms" target ="_blank">https://github.com/caporaso-lab/sample-classifier-ms</a>
**Version:** 2018.6
