### 2018-08-14 a8a3ad1
fix affiliation quotes
