### 2018-08-31 #891 @yochannah
@nbokulich I think this is ready to be reviewed - do you have any suggestions for reviewers?
