### 2018-09-04 #891 @rasbt
I could review the pkg
