### 2018-09-08 #934 @rasbt
@nbokulich @yochannah Just had a quick look at the pdf and the links link all to the manuscript repo rather than the software repo. I remember the issue came up in another review I did some time ago. Best way to address this is to put the manuscript file into the software repo that is being reviewed. E.g., you can put it into the docs subfolder.
