### 2018-09-09 #934 @yochannah
@rasbt Good spot!

@nbokulich could you please put paper.md in the main https://github.com/qiime2/q2-sample-classifier repo? Thanks!
