### 2018-09-10 #934 @nbokulich
sorry for the oversight on my part @yochannah @rasbt

I have added the manuscript to the software repo [here](https://github.com/qiime2/q2-sample-classifier/tree/master/paper)... @yochannah is there some way to reassign the repository to correct the links in the proof? (sorry — I entered that information during submission so do not see a way to edit now)
