### 2018-09-11 #934 @arfon
> @arfon, do we just need to edit the url for the github issue and regenerate the pdf to fix the repo URL, or is there a better way to go about this?

Yep, feel free to edit the issue at the top and then as Whedon to compile again.
