### 2018-09-22 #934 @rasbt
Regarding the checklist above:

>  Repository: Is the source code for this software available at the repository url?

this still has the wrong link, but I think we can ignore it since the paper PDF has the correct link now.
