### 2018-09-22 #934 @rasbt
Regarding

> Version: Does the release version given match the GitHub release (2018.6)?

I see that the softw is at v2018.10 already. I guess the best thing to do is to revisit the version number references at the end, because they version numbers may change during the review stage as potential issues are being addressed
