### 2018-09-22 #934 @rasbt
> Authorship: Has the submitting author (@nbokulich) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

This looks good to me overall, however, I noticed that Benjamin D Kaehler and Gavin A Huttley haven't contributed via GitHub, and there are a few contributors who aren't listed on the paper if I see that correctly. I assume that the authorship has been discussed and handled reasonably and appropriately, but it would be nice, @nbokulich, if you could say a few things about that.
