### 2018-09-22 #934 @rasbt
@nbokulich

regarding

> References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?

Could you please add the DOIs to the paper
