### 2018-09-22 #934 @rasbt
The paper itself looks really great. Just found a few very minor formatting issues -- would be nice if you could address these (ref https://github.com/qiime2/q2-sample-classifier/issues/132)
