### 2018-09-24 #934 @nbokulich
Thank you @rasbt ! I have a few responses, and am working on a PR to address your concerns [here](https://github.com/qiime2/q2-sample-classifier/issues/132) and [here](https://github.com/qiime2/q2-sample-classifier/issues/133) (thank you for raising issues!)

> I see that the softw is at v2018.10 already. I guess the best thing to do is to revisit the version number references at the end, because they version numbers may change during the review stage as potential issues are being addressed

Sorry — we are on a quick release cycle right now. 2018.8 was released a week or two after I submitted this manuscript, and 2018.10 is the current dev version. I agree, we can update later since 2018.10 may be the stable release by the time manuscript review is finalized.

> This looks good to me overall, however, I noticed that Benjamin D Kaehler and Gavin A Huttley haven't contributed via GitHub, and there are a few contributors who aren't listed on the paper if I see that correctly

Thank you for asking. @BenKaehler helped write the paper and contributed to the source code, [this](https://github.com/qiime2/q2-sample-classifier/pull/101) is the most significant contribution (see the commit log) though attributions may have been mixed up since that commit was spliced from a separate PR raised by Ben. @GavinHuttley and @gregcaporaso provided useful discussions, material support, and helped write the manuscript (@gregcaporaso also wrote and reviewed the source code). @thermokarst and @ebolyen both wrote and/or reviewed significant amounts of source code, and helped write the manuscript. @jairideout declined authorship, since he felt that his contributions were minor, and he is acknowledged instead for his support. @q2d2 is a robot. Other contributors to the repo made their contributions after this manuscript was submitted; their work is ongoing and the subject of a separate paper that is in preparation. I would be happy to write up a formal contributions statement in the paper but this did not seem to adhere to journal format.
