### 2018-09-27 #934 @rasbt
@nbokulich

Thanks a lot for the detailed explanation.

> I would be happy to write up a formal contributions statement in the paper but this did not seem to adhere to journal format.

I don't think this is necessary for JOSS, but thanks for asking!!
