### 2018-10-08 #934 @yochannah
Quick status check on this paper... How is everyone getting on?
