### 2018-10-10 #934 @rasbt
@yochannah  @nbokulich  Everything is fine from my side. The examples work and I just ran the unit tests -- except for some warnings in sklearn there are no issues :). So, from my side, it's good to go :)
