### 2018-10-10 #934 @luizirber
I just checked all my boxes. Some comments:
- version: This one I didn't check, but it was already discussed in this thread.
- tests: I didn't find in an easy place how to run the tests, so I ended up downloading the repo and running pytest on it. Is that the appropriate method? Either way, I think it's good to put it somewhere in the docs.
- Congrats on the tutorials and docs! They are really good =]

@yochannah good to go!
