### 2018-10-10 #934 @nbokulich
Thank you @luizirber and @rasbt for your reviews and feedback!

@luizirber — I have issued this PR to address your testing suggestion: https://github.com/qiime2/q2-sample-classifier/pull/138.
