### 2018-10-10 #934 @rasbt
> tests: I didn't find in an easy place how to run the tests, so I ended up downloading the repo and running pytest on it. Is that the appropriate method? Either way, I think it's good to put it somewhere in the docs.

Yes, that's what I did, too. The steps were not mentioned, but one could find them in the travis CI setup file. I agree with @luizirber  that it would not hurt to add a sentence about running the tests with pytest to the Readme file, though.
