### 2018-10-13 #934 @yochannah
Okay, I think the only thing left to do before we ping Arfon to say this is ready is to fix the version numbering and regenerated the paper. @nbokulich - is 2018.10 the correct version to point to?
