### 2018-10-15 #934 @nbokulich
We can just point to version 2018.8 (the current stable release).
