### 2018-10-22 #934 @arfon
@nbokulich - could you update your `references.bib` to be `paper.bib` and update the `paper.md` file accordingly (to point to `paper.md`)?
