### 2018-10-23 #934 @arfon
@nbokulich - actually ignore me. I've just updated the @whedon bot to handle custom bibtex locations (i.e. files not called `paper.bib`)
