### 2018-10-23 #934 @nbokulich
@arfon the "repository" link in the paper PDF points to the correct repo ([here](https://github.com/qiime2/q2-sample-classifier)), but the "software repository" link on the article details page points to the paper repo ([here](https://github.com/caporaso-lab/sample-classifier-ms)) that I originally used at the time of submission (sorry, my fault 😞). Can we get this fixed or has that ship sailed? Thank you!
